requirements.txt
ultimate-notion~=0.7.1
pendulum~=3.0.0
python-dotenv~=1.0.1
google-api-python-client~=2.160.0
google-auth-httplib2~=0.2.0
google-auth-oauthlib~=1.2.1